+++ YyWsnDeviceLibrary/Sensor.cs
@@ -14,14 +14,14 @@ namespace YyWsnDeviceLibrary
         public DateTime SensorTransforTime { get; set; }
 
         public double RSSI { get; set; }
-        public int TXPower { get; set; }
+        public byte TXPower { get; set; }
 
         /// <summary>
         /// 采集和发送的倍数，最小为1，最大为8
         /// </summary>
-        public int TXTimers { get; set; }
+        public byte TXTimers { get; set; }
 
-        public int Frequency { get; set; }
+        public byte Frequency { get; set; }
